moved new custom-env funcs to experimental.h
