+++ quest/src/api/CMakeLists.txt
@@ -5,6 +5,7 @@ target_sources(QuEST
   debug.cpp
   decoherence.cpp
   environment.cpp
+  experimental.cpp
   initialisations.cpp
   matrices.cpp
   modes.cpp
@@ -12,7 +13,6 @@ target_sources(QuEST
   operations.cpp
   paulis.cpp
   qureg.cpp
-  subcommunicator.cpp
   trotterisation.cpp
   types.cpp
 )
